requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.3
plotly==5.15.0
scikit-learn==1.3.0
openai==1.3.0
python-dotenv==1.0.0
altair==5.1.0
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3
requests==2.31.0
pydantic==2.4.2
openpyxl==3.1.2
pydantic==2.4.0
openpyxl==3.1.2
